Port to final Xerces 3.0, conditional use of 64-bit size typedef in some xmlsec APIs, bump version to 1.5.
git-svn-id: https://svn.apache.org/repos/asf/xml/security/trunk@724055 13f79535-47bb-0310-9956-ffa450edef68

+++ c/src/transformers/TXFMC14n.cpp
@@ -164,7 +164,7 @@ unsigned int TXFMC14n::readBytes(XMLByte * const toFill, unsigned int maxToFill)
 
 		return 0;
 
-	return mp_c14n->outputBuffer(toFill, maxToFill);
+	return (unsigned int) mp_c14n->outputBuffer(toFill, maxToFill);
 
 }
